feat: collapse taxonIDs to unique refs per sample
This was done to help reduce duplicate assembly generation resulting
from over-stratification of reads by taxon ID. This appears to be a side
effect of having a kraken2 database with multiple representatives per
viral species. Now, of all taxon IDs mapping to the same reference in
the assembly database, the taxon ID with the highest reads ONLY is
allowed to proceed. Needs more testing.